dialogconversation/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-genai 